+++ src/collective/contact/plonegroup/interfaces.py
@@ -2,6 +2,7 @@
 """Module where all interfaces, events and exceptions live."""
 
 from plone.theme.interfaces import IDefaultPloneLayer
+from Products.PluggableAuthService.interfaces.events import IPASEvent
 from zope.interface import Interface
 
 
@@ -18,4 +19,9 @@ class IPloneGroupContact(Interface):
 class INotPloneGroupContact(Interface):
     """
         Marker interface for non plonegroup contacts.
+    """
+
+
+class IPlonegroupGroupCreatedEvent(IPASEvent):
+    """A new Plone group linked to an organization has been created.
     """